refactor(game): move bulk frontier picker to lib/game/frontier
Follows the single-explore frontier extraction by moving the bulk variant into
the same module:
  - pickFrontierCandidatesBulk (export)
  - collectUnclaimedByRingWalk / collectUnclaimedByMonteCarlo (helpers)

Same clean shape: dependency-injected (`db`), no transaction/write state, all
deps already present in frontier.ts. data-server's `bulkFrontierExploreServer`
orchestrator stays and imports pickFrontierCandidatesBulk. With the pickers gone,
four ./exploration geometry helpers (distanceToNearestOwned,
kingdomRadiusFromCentroid, ringCoords, riskScore) are no longer referenced in
data-server and were dropped from its import.

Behavior-preserving — code moved verbatim. data-server.ts 7,660 → 7,444 lines
(7,865 → 7,444 across the two frontier PRs).

Verified: game + explore suites 1493/1493, tsc + eslint clean.

Stacked on #1569 (frontier pickers) → #1567 (foundation).
